+++ src/views/Player.vue
@@ -230,14 +230,7 @@ export default {
 
       modal.onDidDismiss().then((response) => {
         if (response.data) {
-          const mode = response.data.value.appendToPlaylist
-            ? "append-play"
-            : "replace";
-          apiInstance.post("playlist", {
-            filename: response.data.value.filename,
-            flag: mode,
-            "file-local-options": response.data.value["file-local-options"],
-          });
+          apiInstance.post("playlist", response.data);
         }
       });
       return modal.present();