+++ CHANGELOG.md
@@ -35,3 +35,5 @@ Important Android stuff:
 - Some code tidy-up, optimizations,
 - The app gonna check if the mpv-remote-node version is compatible with the app,
 - [CSS] Improved contrast,
+- Made the whole app more consistent,
+- Set Youtube quality on "Open URL" dialog.